fft_eval: Split SDL and JSON functionality in two programs
The first try to implement JSON output without SDL dependency in commit
0c49c48473ef ("Adds conditional compilation of SDL functionality. Adds
option to output data in JSON format.") still required all dependencies
(including SDL) to be installed for the target system. And it also tried to
compile fft_eval.o against SDL before it then recompiled fft_eval.o again
without SDL.

To avoid these kind of problems, split the fft_eval in three parts:

* fft_eval.c/.h: main code to parse the fft samples
* fft_eval_json.c: json output
* fft_eval_sdl.c: SDL output

The makefile can than build each part separately without the need to
compile some objects twice or to have some unwanted build dependencies.

This build+install target can then be configured using the make parameters:

* CONFIG_fft_eval_sdl=y (or n when it should be disabled)
* CONFIG_fft_eval_json=y (on n when it should be disabled)

The splitup has potential for more cleanup. But this is enough to build
fft_eval_json on OpenWrt without the requirement to have libSDL for the
target system.
